Add bzero & memcpy

+++ ft_memset.c
@@ -10,11 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
+#include "libft.h"
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	unsigned long	i;
+	unsigned int	i;
 	char			*str;
 
 	str = (char *) b;
